separate Encoder App to a simple template
