+++ .gitignore
@@ -7,3 +7,5 @@ build
 .docusaurus
 .docusaurus/
 .docusaurus/*
+
+.env
