+++ src/components/EncoderAppPage/EncNode.js
