--- src/pages/medical-term-coding/meddra.js
@@ -1,18 +1,0 @@
-import React from "react";
-
-import "../index.css";
-
-import NavBar from "../../components/NavBar";
-import MainSection from "../../components/MainSection";
-import Features from "../../components/Features";
-import UseCases from "../../components/UseCases";
-import GettingStarted from "../../components/GettingStarted";
-import Footer from "../../components/Footer";
-
-export default function EncoderPage() {
-  return (
-    <div>
-      this is an encoder page
-    </div>
-  );
-}
